+++ examples/sedk-neo4j-ts-example/index.ts
@@ -1,0 +1,19 @@
+async function importReadmeExample() {
+	return await import('./readme-example.js')
+}
+
+const importFunctionsList = [
+	importReadmeExample,
+]
+
+async function main() {
+	for (const fn of importFunctionsList) {
+		console.log(`📦 run ${fn.name}`)
+		await fn()
+	}
+}
+
+console.log('🚀 Starting sedk-postgres-js-example ...')
+main().then(() => {
+	console.log('👋 End of sedk-postgres-js-example')
+})