+++ packages/sedk-neo4j/src/Builder.ts
@@ -1,0 +1,5 @@
+import { RootStep } from './steps/RootStep'
+
+export function builder(): RootStep {
+	return new RootStep()
+}