+++ packages/sedk-neo4j/src/steps/types.ts
@@ -1,0 +1,6 @@
+import { Variable } from '../Variable'
+import { Label } from '../Label'
+import { Asterisk } from '../singletoneConstants'
+
+export type VarLabels = [(Variable | Label), ...Label[]]
+export type ReturnItems = Variable[] | [...Variable[], Asterisk]