+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.1.2]  2023-01-05
+### Added
+- add package sedk-neo4j to workspace
+- add example project sedk-neo4j-ts-example
+
 ## [0.0.1]  2023-01-05
 ### Added
 - workspace for  sedk-postgres